+++ NHS111/NHS111.Web.Test/Builders/DOSBuilderTests.cs
@@ -186,12 +186,12 @@ namespace NHS111.Web.Presentation.Builders.Tests
         }
 
 
-        private bool AssertIsMetric(HttpRequestMessage request, double original) {
+        private bool AssertIsMetric(HttpRequestMessage request, int original) {
             var content = request.Content.ReadAsStringAsync().Result;
             var payload = JsonConvert.DeserializeObject<DosCase>(content);
 
             const float MILES_PER_KM = 1.609344f;
-            return payload.SearchDistance == Math.Ceiling(original / MILES_PER_KM);
+            return payload.SearchDistance == (int)Math.Ceiling(original / MILES_PER_KM);
         }
 
         private void MockRestfulHelperWithExpectedUrl(string expectedSymptomGroup)
